ff-parsers/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ff-parsers"
version = "1.0.0"
description = "Lightweight document parsing library for extracting content from various file formats"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "Fenixflow", email = "dev@fenixflow.com"}
]
dependencies = [
    "pdfplumber>=0.10.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.0.0",
    "pytesseract>=0.3.0",
    "pillow>=10.0.0",
    "chardet>=5.0.0",
    "python-magic>=0.4.0",
    "python-magic-bin>=0.4.14; platform_system=='Windows'",
    "pydantic>=2.4,<3.0",
    "markdown-it-py>=3.0,<4.0",
    "markitdown[all]>=0.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
    "pytest-cov>=4.0",
]

[project.urls]
repository = "https://gitlab.com/fenixflow/fenix-packages"
documentation = "https://gitlab.com/fenixflow/fenix-packages/-/tree/main/ff-parsers"

[project.scripts]
ff-ingest = "ff_parsers.ingest:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
